requirements.txt
Flask==2.2.5
gspread==5.12.0
google-auth==2.22.0
google-api-python-client==2.104.0
email-validator==2.1.0.post1
